add better logging of errors while parsing json input
this will now print the data received from pnpm/stdin if an error occurs
in addition to the zod error (formatted using `.format()`, which
somewhat simplifies the zod error but looses a bit of information in the
process). A few default zod messages are now overwritten to better
reflect what is actually relevant in the input data.

Additionally we now check for pnpm errors and treat them differently
from zod validation errors